rag_service/requirements.txt
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

